Add ErrorHandlingMiddleware

+++ WasteSegregation.Application/Services/RealEstateService.cs
@@ -4,27 +4,24 @@ public class RealEstateService : IRealEstateService
 {
     private readonly IRealEstateRepository realEstateRepository;
     private readonly IMapper mapper;
-    private readonly ILogger<RealEstateService> logger;
-
-    public RealEstateService(IRealEstateRepository realEstateRepository, IMapper mapper, ILogger<RealEstateService> logger)
+    
+    public RealEstateService(IRealEstateRepository realEstateRepository, IMapper mapper)
     {
         this.realEstateRepository = realEstateRepository;
         this.mapper = mapper;
-        this.logger = logger;
     }
 
     public async Task<IEnumerable<RealEstateDto>> GetAllAsync()
     {
-        logger.LogInformation("Get all real estates");
         var realEstate = await realEstateRepository.GetAllAsync();
-        if (realEstate == null) return null;
+        if (realEstate == null) throw new NotFoundException("Real estate not found");
         return mapper.Map<List<RealEstateDto>>(realEstate);
     }
 
     public async Task<RealEstateDto> GetByIdAsync(int id)
     {
         var realEstate = await realEstateRepository.GetByIdAsync(id);
-        if (realEstate == null) return null;
+        if (realEstate == null) throw new NotFoundException("Real estate not found");
         return mapper.Map<RealEstateDto>(realEstate);
     }
 
@@ -38,14 +35,15 @@ public class RealEstateService : IRealEstateService
     public async Task UpdateAsync(UpdateRealEstateDto updateRealEstate, int id)
     {
         var existingRealEstate = await realEstateRepository.GetByIdAsync(id);
+        if (existingRealEstate == null) throw new NotFoundException("Real estate not found");
         var realEstate = mapper.Map(updateRealEstate, existingRealEstate);
         await realEstateRepository.UpdateAsync(realEstate);
     }
 
     public async Task DeleteAsync(int id)
     {
-        logger.LogError($"Error when delete real estate with id: {id}");
         var realEstate = await realEstateRepository.GetByIdAsync(id);
+        if (realEstate == null) throw new NotFoundException("Real estate not found");
         await realEstateRepository.DeleteAsync(realEstate);
     }
 }
